lesson-2: refactoring packages
